Add description to README on how to set up an S3 bucket for BDBag w/ expiration. (#47)
Adds a paragraph to the README describing in a step-by-step manner how to set up an AWS S3 bucket for BDBag. The special feature of this bucket is that it has a 1-day lifecycle expiration. This is appropriate as it economic and also contains the risk of data abuse. Adding the expiration feature is also described in a step-by-step list.