Adding environment variables for persistent BDBag in S3 (#49) (resolves #54)
This is part of storing a BDBag in AWS-S3 for a limited time. 

I added prompt question to `install_bootstrap` re. S3 bucket name, which asks user for the name of the AWS S3 bucket BDBags will be stored in. In addition, I added the environment variables AZUL_S3_BUCKET, AWS_ACCESS_KEY and AWS_SECRET_ACCESS_KEY
to the following files

* `install_bootstrap`
* `boardwalk/dev.yml`
* `boardwalk/prod.yml`
* `boardwalk/conf/boardwalk.config.template`

so they can be read into container `dcc_dashboard-service_1`. To the `README` file I

* Added a reminder for having AWS credentials ready to section `Collecting Information`.
* Added a reminder to provide exact name of S3 bucket in the install script.