Changed position of region question and added default expiration time
The question for the default AWS region of the azul-s3-bucket now follows now
right on the heels of the question for the name of the azul-s3-bucket. That way
the significance of that question should become more obvious.

I assigned 3600 s as the default value for the presigned URL of the BDBag
in the azul-s3-bucket.