Simplified and adapted docker compose files
As part of removing the `dev` option when configuring the nginx
proxy server I removed the Docker compose configuration files
`dev.yml` and `prod.yml` in directory cgp-deployment/common.

Next, I renamed the file `base.yml` to `docker-compose.yml`. That
way a user can execute `docker-compose up` or `docker-compose down`
in cgp-deployment/common, instead of the more complicated command
that was required prior to that change.