Adjusted text in install_bootstrap
Delete lines which referred to the choice of selecting
nginx in dev or prod mode, and removed the corresponding
bash variable.

Now with only one Docker compose configuration file in
the common directory, I adjusted the command in the
script to reflect that to bring up and down the
docker-compose pool of containers.