Remove nginx dev option for common containers (resolves #82)
We currently do not see a need to have a dev deployment for
the _Commons_ Dashboard/Boardwalk. Such a dev
deployment had installed "fake" certificates. Specifically, the
following changes were made:
* Removed choice question of nginx common dev and prod install.
* Simplified and adapted docker compose files.
* As part of removing the `dev` option when configuring the nginx
proxy server I removed the Docker compose configuration files
`dev.yml` and `prod.yml` in directory cgp-deployment/common.
* Renamed the file `base.yml` to `docker-compose.yml`. That
way a user can execute `docker-compose up` or `docker-compose down`
in cgp-deployment/common, instead of the more complicated command
that was required prior to that change.
* Adjusted text in install_bootstrap: deleted lines which referred to
the choice of selecting nginx in dev or prod mode, and removed
the corresponding bash variable.
* Finally, with only one Docker compose configuration file in
the common directory, I adjusted the command in the script to
reflect that to bring up and down the docker-compose pool of containers.